refactor: Move some db functionality
