+++ airbyte-integrations/connectors/source-firebolt/source_firebolt/utils.py
@@ -2,42 +2,11 @@
 # Copyright (c) 2021 Airbyte, Inc., all rights reserved.
 #
 
-import json
 from datetime import date, datetime
 from decimal import Decimal
 from typing import Any, Dict, List, Optional, Union
 
-from airbyte_cdk.logger import AirbyteLogger
 from airbyte_cdk.models import AirbyteMessage, AirbyteRecordMessage, Type
-from firebolt.client import DEFAULT_API_URL
-
-
-def parse_config(config: json, logger: AirbyteLogger) -> Dict[str, Any]:
-    """
-    Convert dict of config values to firebolt.db.Connection arguments
-
-    :param config: json-compatible dict of settings
-    :param logger: AirbyteLogger instance to print logs.
-
-    :return: dictionary of firebolt.db.Connection-compatible kwargs
-    """
-    connection_args = {
-        "database": config["database"],
-        "username": config["username"],
-        "password": config["password"],
-        "api_endpoint": config.get("host", DEFAULT_API_URL),
-        "account_name": config.get("account"),
-    }
-    # engine can be a name or a full URL of a cluster
-    engine = config.get("engine")
-    if engine:
-        if "." in engine:
-            connection_args["engine_url"] = engine
-        else:
-            connection_args["engine_name"] = engine
-    else:
-        logger.info("Engine parameter was not provided. Connecting to the default engine.")
-    return connection_args
 
 
 def convert_type(fb_type: str, nullable: bool) -> Dict[str, Union[str, Dict]]:
